added general tests

+++ contracts/Farm.sol
@@ -2,10 +2,10 @@
 
 pragma solidity 0.6.12;
 
-import '@pancakeswap/pancake-swap-lib/contracts/math/SafeMath.sol';
-import '@pancakeswap/pancake-swap-lib/contracts/token/BEP20/IBEP20.sol';
-import '@pancakeswap/pancake-swap-lib/contracts/token/BEP20/SafeBEP20.sol';
-import '@pancakeswap/pancake-swap-lib/contracts/access/Ownable.sol';
+import "@pancakeswap/pancake-swap-lib/contracts/math/SafeMath.sol";
+import "@pancakeswap/pancake-swap-lib/contracts/token/BEP20/IBEP20.sol";
+import "@pancakeswap/pancake-swap-lib/contracts/token/BEP20/SafeBEP20.sol";
+import "@pancakeswap/pancake-swap-lib/contracts/access/Ownable.sol";
 
 import "./MasterPredWallet.sol";
 
@@ -37,7 +37,7 @@ contract MasterPred is Ownable {
 
     // Info of each user.
     struct UserInfo {
-        uint256 amount;     // How many LP tokens the user has provided.
+        uint256 amount; // How many LP tokens the user has provided.
         uint256 rewardDebt; // Reward debt. See explanation below.
         //
         // We do some fancy math here. Basically, any point in time, the amount of PREDs
@@ -54,9 +54,9 @@ contract MasterPred is Ownable {
 
     // Info of each pool.
     struct PoolInfo {
-        IBEP20 lpToken;           // Address of LP token contract.
-        uint256 allocPoint;       // How many allocation points assigned to this pool. PREDs to distribute per block.
-        uint256 lastRewardBlock;  // Last block number that PREDs distribution occurs.
+        IBEP20 lpToken; // Address of LP token contract.
+        uint256 allocPoint; // How many allocation points assigned to this pool. PREDs to distribute per block.
+        uint256 lastRewardBlock; // Last block number that PREDs distribution occurs.
         uint256 accPredPerShare; // Accumulated PREDs per share, times 1e12. See below.
     }
 
@@ -74,7 +74,7 @@ contract MasterPred is Ownable {
     // Info of each pool.
     PoolInfo[] public poolInfo;
     // Info of each user that stakes LP tokens.
-    mapping (uint256 => mapping (address => UserInfo)) public userInfo;
+    mapping(uint256 => mapping(address => UserInfo)) public userInfo;
     // Total allocation points. Must be the sum of all allocation points in all pools.
     uint256 public totalAllocPoint = 0;
     // The block number when PRED mining starts.
@@ -84,7 +84,11 @@ contract MasterPred is Ownable {
 
     event Deposit(address indexed user, uint256 indexed pid, uint256 amount);
     event Withdraw(address indexed user, uint256 indexed pid, uint256 amount);
-    event EmergencyWithdraw(address indexed user, uint256 indexed pid, uint256 amount);
+    event EmergencyWithdraw(
+        address indexed user,
+        uint256 indexed pid,
+        uint256 amount
+    );
 
     constructor(
         IBEP20 _pred,
@@ -95,14 +99,16 @@ contract MasterPred is Ownable {
         predPerBlock = _predPerBlock;
         startBlock = _startBlock;
         wallet = new MasterPredWallet(_pred);
-        
+
         // staking pool
-        poolInfo.push(PoolInfo({
-            lpToken: _pred,
-            allocPoint: 200,
-            lastRewardBlock: startBlock,
-            accPredPerShare: 0
-        }));
+        poolInfo.push(
+            PoolInfo({
+                lpToken: _pred,
+                allocPoint: 200,
+                lastRewardBlock: startBlock,
+                accPredPerShare: 0
+            })
+        );
 
         totalAllocPoint = 200;
     }
@@ -117,22 +123,34 @@ contract MasterPred is Ownable {
 
     // Add a new lp to the pool. Can only be called by the owner.
     // XXX DO NOT add the same LP token more than once. Rewards will be messed up if you do.
-    function add(uint256 _allocPoint, IBEP20 _lpToken, bool _withUpdate) public onlyOwner {
+    function add(
+        uint256 _allocPoint,
+        IBEP20 _lpToken,
+        bool _withUpdate
+    ) public onlyOwner {
         if (_withUpdate) {
             massUpdatePools();
         }
-        uint256 lastRewardBlock = block.number > startBlock ? block.number : startBlock;
+        uint256 lastRewardBlock = block.number > startBlock
+            ? block.number
+            : startBlock;
         totalAllocPoint = totalAllocPoint.add(_allocPoint);
-        poolInfo.push(PoolInfo({
-            lpToken: _lpToken,
-            allocPoint: _allocPoint,
-            lastRewardBlock: lastRewardBlock,
-            accPredPerShare: 0
-        }));
+        poolInfo.push(
+            PoolInfo({
+                lpToken: _lpToken,
+                allocPoint: _allocPoint,
+                lastRewardBlock: lastRewardBlock,
+                accPredPerShare: 0
+            })
+        );
     }
 
     // Update the given pool's PRED allocation point. Can only be called by the owner.
-    function set(uint256 _pid, uint256 _allocPoint, bool _withUpdate) public onlyOwner {
+    function set(
+        uint256 _pid,
+        uint256 _allocPoint,
+        bool _withUpdate
+    ) public onlyOwner {
         if (_withUpdate) {
             massUpdatePools();
         }
@@ -140,7 +158,9 @@ contract MasterPred is Ownable {
         uint256 prevAllocPoint = poolInfo[_pid].allocPoint;
         poolInfo[_pid].allocPoint = _allocPoint;
         if (prevAllocPoint != _allocPoint) {
-            totalAllocPoint = totalAllocPoint.sub(prevAllocPoint).add(_allocPoint);
+            totalAllocPoint = totalAllocPoint.sub(prevAllocPoint).add(
+                _allocPoint
+            );
         }
     }
 
@@ -162,24 +182,42 @@ contract MasterPred is Ownable {
     }
 
     // Return reward multiplier over the given _from to _to block.
-    function getMultiplier(uint256 _from, uint256 _to) public view returns (uint256) {
+    function getMultiplier(uint256 _from, uint256 _to)
+        public
+        view
+        returns (uint256)
+    {
         return _to.sub(_from).mul(BONUS_MULTIPLIER);
     }
 
     // View function to see pending PREDs on frontend.
-    function pendingPred(uint256 _pid, address _user) external view returns (uint256) {
+    function pendingPred(uint256 _pid, address _user)
+        external
+        view
+        returns (uint256)
+    {
         PoolInfo storage pool = poolInfo[_pid];
         UserInfo storage user = userInfo[_pid][_user];
         uint256 accPredPerShare = pool.accPredPerShare;
         uint256 lpSupply = pool.lpToken.balanceOf(address(this));
         if (block.number > pool.lastRewardBlock && lpSupply != 0) {
-            uint256 multiplier = getMultiplier(pool.lastRewardBlock, block.number);
-            uint256 predReward = multiplier.mul(predPerBlock).mul(pool.allocPoint).div(totalAllocPoint);
-            uint256 predBal = pred.balanceOf(address(wallet)).sub(totalRewardDebt);
+            uint256 multiplier = getMultiplier(
+                pool.lastRewardBlock,
+                block.number
+            );
+            uint256 predReward = multiplier
+            .mul(predPerBlock)
+            .mul(pool.allocPoint)
+            .div(totalAllocPoint);
+            uint256 predBal = pred.balanceOf(address(wallet)).sub(
+                totalRewardDebt
+            );
             if (predReward >= predBal) {
-                predReward = predBal;        
+                predReward = predBal;
             }
-            accPredPerShare = accPredPerShare.add(predReward.mul(1e12).div(lpSupply));    
+            accPredPerShare = accPredPerShare.add(
+                predReward.mul(1e12).div(lpSupply)
+            );
         }
         return user.amount.mul(accPredPerShare).div(1e12).sub(user.rewardDebt);
     }
@@ -191,7 +229,6 @@ contract MasterPred is Ownable {
             updatePool(pid);
         }
     }
-
 
     // Update reward variables of the given pool to be up-to-date.
     function updatePool(uint256 _pid) public {
@@ -205,13 +242,18 @@ contract MasterPred is Ownable {
             return;
         }
         uint256 multiplier = getMultiplier(pool.lastRewardBlock, block.number);
-        uint256 predReward = multiplier.mul(predPerBlock).mul(pool.allocPoint).div(totalAllocPoint);
+        uint256 predReward = multiplier
+        .mul(predPerBlock)
+        .mul(pool.allocPoint)
+        .div(totalAllocPoint);
         uint256 predBal = pred.balanceOf(address(wallet)).sub(totalRewardDebt);
         if (predReward >= predBal) {
-            predReward = predBal;        
+            predReward = predBal;
         }
-        
-        pool.accPredPerShare = pool.accPredPerShare.add(predReward.mul(1e12).div(lpSupply));
+
+        pool.accPredPerShare = pool.accPredPerShare.add(
+            predReward.mul(1e12).div(lpSupply)
+        );
         totalRewardDebt = totalRewardDebt.add(predReward);
         pool.lastRewardBlock = block.number;
     }
@@ -222,15 +264,25 @@ contract MasterPred is Ownable {
         UserInfo storage user = userInfo[_pid][msg.sender];
         updatePool(_pid);
         if (user.amount > 0) {
-            uint256 pending = user.amount.mul(pool.accPredPerShare).div(1e12).sub(user.rewardDebt);
-            if(pending > 0) {
+            uint256 pending = user
+            .amount
+            .mul(pool.accPredPerShare)
+            .div(1e12)
+            .sub(user.rewardDebt);
+            if (pending > 0) {
                 safePredTransfer(msg.sender, pending);
             }
         }
         if (_amount > 0) {
-            uint balBefore = pool.lpToken.balanceOf(address(this));
-            pool.lpToken.safeTransferFrom(address(msg.sender), address(this), _amount);
-            user.amount = user.amount.add(pool.lpToken.balanceOf(address(this)).sub(balBefore));
+            uint256 balBefore = pool.lpToken.balanceOf(address(this));
+            pool.lpToken.safeTransferFrom(
+                address(msg.sender),
+                address(this),
+                _amount
+            );
+            user.amount = user.amount.add(
+                pool.lpToken.balanceOf(address(this)).sub(balBefore)
+            );
         }
         user.rewardDebt = user.amount.mul(pool.accPredPerShare).div(1e12);
         emit Deposit(msg.sender, _pid, _amount);
@@ -243,18 +295,19 @@ contract MasterPred is Ownable {
         require(user.amount >= _amount, "withdraw: not good");
 
         updatePool(_pid);
-        uint256 pending = user.amount.mul(pool.accPredPerShare).div(1e12).sub(user.rewardDebt);
-        if(pending > 0) {
+        uint256 pending = user.amount.mul(pool.accPredPerShare).div(1e12).sub(
+            user.rewardDebt
+        );
+        if (pending > 0) {
             safePredTransfer(msg.sender, pending);
         }
-        if(_amount > 0) {
+        if (_amount > 0) {
             user.amount = user.amount.sub(_amount);
             pool.lpToken.safeTransfer(address(msg.sender), _amount);
         }
         user.rewardDebt = user.amount.mul(pool.accPredPerShare).div(1e12);
         emit Withdraw(msg.sender, _pid, _amount);
     }
-
 
     // Withdraw without caring about rewards. EMERGENCY ONLY.
     function emergencyWithdraw(uint256 _pid) public {
@@ -268,7 +321,8 @@ contract MasterPred is Ownable {
 
     // Safe pred transfer function, just in case if rounding error causes pool to not have enough PREDs.
     function safePredTransfer(address _to, uint256 _amount) internal {
-        totalRewardDebt = totalRewardDebt.sub(wallet.safePredTransfer(_to, _amount));
-        
+        totalRewardDebt = totalRewardDebt.sub(
+            wallet.safePredTransfer(_to, _amount)
+        );
     }
 }
